Enable --daemon flag
Removed pilifanad command and enabled --daemon flag for pilifana command
to run Pilifana as daemon instead of a console executable.
